feat: Add AWS RDS database and EC2 instance resources
This commit adds the necessary Terraform code to create an AWS RDS database instance and an EC2 instance. The RDS instance is configured with MySQL engine version 8.0 and a micro instance class. The EC2 instance is based on the specified AMI and is configured with Apache HTTP server, PHP, and MySQL. The user data script installs necessary packages, starts the HTTP server, and creates a basic PHP info page.

Closes #123